refactor: split solvent runtime by context
